// admin-frontend/src/api/generated/model/menuVO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Admin Backend API
 * 后台管理系统后端 API 文档，前端可通过 OpenAPI Schema 自动生成接口调用代码
 * OpenAPI spec version: 1.0.0
 */

/**
 * 菜单树形视图
 */
export interface MenuVO {
  /** 菜单ID */
  id?: number;
  /** 菜单名称 */
  name?: string;
  /** 路由路径 */
  path?: string;
  /** 菜单图标 */
  icon?: string;
  /** 前端组件路径 */
  component?: string;
  /** 父级ID */
  parentId?: number;
  /** 菜单类型（directory-目录 menu-菜单 button-按钮） */
  type?: string;
  /** 排序 */
  sort?: number;
  /** 子菜单列表 */
  children?: MenuVO[];
}